+++ E2E-Integration-Test/src/test/java/ru/rosatom/e2e/userTests/P1000_Authorization/AdminAuthRequest.java
@@ -5,7 +5,7 @@ import ru.rosatom.e2e.user.UserAuthorizationRequest;
 public class AdminAuthRequest  extends UserAuthorizationRequest {
 
     public AdminAuthRequest() {
-        super("admin@mail.ru", "251323Nn");
+        super("admin@greenatom.ru", "251323Nn");
 
     }
 }
