+++ E2E-Integration-Test/src/test/java/ru/rosatom/e2e/userTests/P1500_Organizations/OrganizationSearchRequestId.java
@@ -1,4 +1,4 @@
-package ru.rosatom.e2e.organizationTest.P1000_Organization;
+package ru.rosatom.e2e.userTests.P1500_Organizations;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Data
 @AllArgsConstructor
-public class OrganizationSearchRequestName {
-    private String name;
+public class OrganizationSearchRequestId {
+    private Integer id;
 
 }
